+++ StardewArchipelago/Locations/CodeInjections/Modded/ModdedEventInjections.cs
@@ -5,12 +5,21 @@ using Microsoft.Xna.Framework;
 using StardewModdingAPI;
 using StardewArchipelago.Archipelago;
 using StardewValley;
+using StardewArchipelago.Locations.CodeInjections.Modded.SVE;
+using System.Diagnostics.Tracing;
+using Archipelago.MultiClient.Net.Packets;
 
 namespace StardewArchipelago.Locations.CodeInjections.Modded
 {
     public static class ModdedEventInjections
     {
+        private static readonly Dictionary<string, string[]> Base_Events = new(){
 
+        };
+        private static Dictionary<string, string[]> Total_Events = new(){
+
+        };
+        private const string DUMMY_EVENT_DATA = "continue/-200 -200 0/Lewis -10 -10 0/pause 50/end";
         private static readonly string RECIPE_SUFFIX = " Recipe";
         private static readonly Dictionary<int, string> eventCooking = new(){
             {181091234, "Mushroom Kebab"},
@@ -37,8 +46,70 @@ namespace StardewArchipelago.Locations.CodeInjections.Modded
             _modHelper = modHelper;
             _archipelago = archipelago;
             _locationChecker = locationChecker;
+            GenerateEventKeys();
+        }
+        
+        public static void InitializeSVEEvents()
+        {
+            ReplaceViewableCutscenes();
+            ReplaceLockedCutscenes();
         }
 
+        // Some events we want to see, so this will let them play out, so we make up AP event scene requirements.
+        public static void ReplaceViewableCutscenes()
+        {
+            var farmEvents = Game1.content.Load<Dictionary<string, string>>("Data\\Events\\Farm");
+            var farmhouseEvents = Game1.content.Load<Dictionary<string, string>>("Data\\Events\\FarmHouse");
+            SVECutsceneInjections.SVE_ReplaceViewableCutscenes(farmEvents, farmhouseEvents);
+        }
+
+        public static void GenerateEventKeys()
+        {
+            foreach (var (eventMapName, eventKeys) in Base_Events)
+            {
+                Total_Events[eventMapName] = eventKeys;
+            }
+            foreach (var (eventMapName, eventKeys) in SVECutsceneInjections.SVE_Events)
+            {
+                Total_Events[eventMapName] = eventKeys;
+            }
+        }
+
+        // Otherwise, we should just lock the events out of being seen at all and simply toggled by eventSeen.
+        public static void ReplaceLockedCutscenes()
+        {
+            foreach (var (eventMapName, eventKeys) in Total_Events)
+            {
+                var mapName = eventMapName;
+                if (mapName.Contains("FarmHouse"))
+                {
+                    mapName = "FarmHouse";
+                }
+
+                if (mapName.Contains("Custom_AdventurerSummit"))
+                {
+                    mapName = "Custom_AdventurerSummit";
+                }
+
+                var currentMapEventData = Game1.content.Load<Dictionary<string, string>>("Data\\Events\\" + mapName);
+                foreach (var eventKey in eventKeys)
+                {
+                    var eventID = eventKey.Split("/")[0];
+                    // Sometimes, CP has not added the cutscene to the event list yet due to its dynamic adding methods...
+                    if (!currentMapEventData.ContainsKey(eventKey))
+                    {
+                        // ... so just add a dummy event if that's the case to make sure its locked and move on to the next event.
+                        currentMapEventData[eventKey] = DUMMY_EVENT_DATA;
+                    }
+
+                    // Everything exists now, so just make the requirement for the event itself.
+                    var newEventKey = eventKey + "/e " + eventID;
+                    var eventData = currentMapEventData[eventKey];
+                    currentMapEventData.Remove(eventKey);
+                    currentMapEventData[newEventKey] = eventData;
+                }
+            }
+        }
         public static bool AddCookingRecipe_CheckForStrayRecipe_Prefix(Event __instance, GameLocation location, GameTime time, string[] split)
         {
             try
